Added Custom Inspector for DictionarySO. Modifying a dictionary in a DictionarySO asset is now much more comfortable and visually intuitive.
There's an issue with SerializedProperty of Array type - if you try to delete an item in an array at a certain index, then if the item type can be nullable and the value is not null, it will get nulled instead and will be deleted only after the second attempt. So there's an editor extension in the EditorExtensions file, a method called SetPropertyValueToNullIfPossible, which nulls the value if it is possible. But currently the only type it actually nulls is 'ObjectReference'. So it needs to be completed for all the cases in the switch statement, or it might not work.